Update the modal to the specification outlined
Divided the modal into different sections to allow the use of flexbox.

Made the "!" icon its own section so it could stay to the left of the
rest of the content.

Emboldened the header font as well as made the size bigger.

Made the header and the close icon one division as to say aligned with
one another.

Made the continue and cancel buttons one division so they could be
aligned with one another.

Added padding within the modal.

Put gaps between the different divisions to space out the content.

This is probably the worst commit ever since it is not atomic at all but
I got WAY into trying to figure flexbox out in this one without
committing anything lol.