Update header format according to specifications
Created a style selection for the header class in the CSS style sheet.

Chagned the display to flex so that all the items in the header would be
aligned across one axis.

Split the two link section from one another so that they are separate
from one another across the page.

Added some padding to the header as well.