Modify footer format to specifications
Create style settings in CSS sheet for footer class.

Separated the footer link components across from one another on the page.

Added padding and a background to the footer.

Updated the display of the body component to flex, changed the direction
to column, and justified all items as space between so that the footer
was at the bottom of the page.